@@ -1,7 +1,7 @@
 {
   "name": "@karamem0/zenn-plugin-emoji",
   "version": "0.0.0",
-  "type": "module",
+  "type": "commonjs",
   "main": "dist/index.js",
   "bin": {
     "zenn-emoji": "dist/index.js"
@@ -19,7 +19,7 @@
     "zenn"
   ],
   "scripts": {
-    "build": "babel src --extensions '.ts' --out-dir dist --copy-files",
+    "build": "tsc && webpack",
     "eslint": "eslint src",
     "eslint:fix": "eslint src --fix"
   },
@@ -27,19 +27,24 @@
     "@azure/identity": "^4.10.2",
     "commander": "^14.0.0",
     "dotenv": "^17.2.0",
-    "openai": "^5.8.3"
+    "glob": "^11.0.3",
+    "openai": "^5.10.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.28.0",
     "@babel/preset-typescript": "^7.27.1",
-    "@eslint/js": "^9.30.1",
-    "@stylistic/eslint-plugin": "^5.1.0",
-    "@types/node": "^24.0.12",
-    "babel-preset-minify": "^0.5.2",
-    "eslint": "^9.30.1",
+    "@eslint/js": "^9.31.0",
+    "@stylistic/eslint-plugin": "^5.2.0",
+    "@types/node": "^24.0.14",
+    "babel-loader": "^10.0.0",
+    "copy-webpack-plugin": "^13.0.0",
+    "eslint": "^9.31.0",
     "globals": "^16.3.0",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.36.0"
+    "typescript-eslint": "^8.37.0",
+    "webpack": "^5.100.2",
+    "webpack-cli": "^6.0.1",
+    "webpack-node-externals": "^3.0.0"
   },
   "volta": {
     "node": "22.17.0"
